change library_type on dataset page, fix publication link
